+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/geometry/Circle.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.test;
+package com.rakovets.course.java.core.practice.oop_principles.geometry;
 
 public class Circle extends Figure implements Length {
     private final double radius;
