+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/geometry/Figure.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.test;
+package com.rakovets.course.java.core.practice.oop_principles.geometry;
 
 public abstract class Figure implements MainUtils {
     @Override
