+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/geometry/Perimeter.java
@@ -1,0 +1,5 @@
+package com.rakovets.course.java.core.practice.oop_principles.geometry;
+
+public interface Perimeter {
+    double perimeter();
+}
