+++ src/main/java/com/rakovets/course/java/core/practice/oop_principles/geometry/Triangle.java
@@ -1,4 +1,4 @@
-package com.rakovets.course.java.core.practice.oop_principles.test;
+package com.rakovets.course.java.core.practice.oop_principles.geometry;
 
 public class Triangle extends Figure implements Height, Perimeter {
     private final double sideA;
