--- src/main/java/com/rakovets/course/java/core/practice/oop_principles/test/Height.java
@@ -1,5 +1,0 @@
-package com.rakovets.course.java.core.practice.oop_principles.test;
-
-public interface Height {
-    double height();
-}
