--- src/main/java/com/rakovets/course/java/core/practice/oop_principles/test/Perimeter.java
@@ -1,5 +1,0 @@
-package com.rakovets.course.java.core.practice.oop_principles.test;
-
-public interface Perimeter {
-    double perimeter();
-}
